+++ packages/reporter/cypress/plugins/index.ts
@@ -1,1 +1,11 @@
-export default () => {}
+const cp = require('child_process')
+const percyHealthCheck = require('@percy/cypress/task')
+
+cp.exec('http-server -p 5006 dist')
+
+/**
+ * @type {Cypress.PluginConfig}
+ */
+module.exports = (on: Function) => {
+  on('task', percyHealthCheck)
+}