+++ packages/server/lib/errors.js
@@ -903,8 +903,8 @@ const getMsgByType = function (type, arg1 = {}, arg2, arg3) {
         You can safely remove this option from your config.`
     case 'EXPERIMENTAL_NETWORK_STUBBING_REMOVED':
       return stripIndent`\
-        The \`experimentalNetworkStubbing\` configuration option was removed in Cypress version \`6.0.0\`. 
-        It is no longer necessary for using \`cy.http()\` (formerly \`cy.route2()\`).
+        The \`experimentalNetworkStubbing\` configuration option was removed in Cypress version \`6.0.0\`.
+        It is no longer necessary for using \`cy.intercept()\` (formerly \`cy.route2()\`).
 
         You can safely remove this option from your config.`
     case 'INCOMPATIBLE_PLUGIN_RETRIES':
@@ -920,7 +920,7 @@ const getMsgByType = function (type, arg1 = {}, arg2, arg3) {
     case 'INVALID_CONFIG_OPTION':
       return stripIndent`\
         ${arg1.map((arg) => `\`${arg}\` is not a valid configuration option`)}
-        
+
         https://on.cypress.io/configuration
         `
     default: